git-svn-id: svn://123.57.173.36/meijia-parent@681 52ac4245-7803-4e87-b018-04b7c5ce731e

+++ simi-app/src/main/java/com/simi/action/app/user/UserController.java
@@ -334,7 +334,7 @@ public class UserController extends BaseController {
 			@RequestParam("name") String name,
 			@RequestParam("realName") String realName,
 			@RequestParam("sex") String sex,
-			@RequestParam("degreeId") Short degreeId,
+			@RequestParam("degreeId") Long degreeId,
 			@RequestParam("major") String major,
 			@RequestParam("idCard") String idCard,
 			@RequestParam(value = "tag_ids", required = false, defaultValue = "") String tagIds) {
